fix: rest of the types
