+++ packages/zustand-x/src/utils/generateStateGetSelectors.ts
@@ -1,4 +1,4 @@
-import { TGetSelectorRecord, TReadonlyStoreApi } from './types.v2';
+import { TGetSelectorRecord, TReadonlyStoreApi } from '../types';
 
 export const generateStateGetSelectors = <T>(store: TReadonlyStoreApi<T>) => {
   const selectors: TGetSelectorRecord<T> = {} as TGetSelectorRecord<T>;
